nit: [Compilation] move all from Compilation1.java to CompilationImpl.java
